wizard status and menu updates
- updated deletePositionRequest endpoint to only allow deletion of draft position requests that belong to currently logged in user
- refactored status indicator into it's own component, added to wizard header

+++ apps/api/src/modules/position-request/position-request.service.ts
@@ -1001,8 +1001,8 @@ export class PositionRequestApiService {
     }
   }
 
-  async deletePositionRequest(id: number) {
-    const result = await this.prisma.positionRequest.delete({ where: { id } });
+  async deletePositionRequest(id: number, userId: string) {
+    const result = await this.prisma.positionRequest.delete({ where: { id, user_id: userId, status: 'DRAFT' } });
     return result;
   }
